services/sap-b1-adapter-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
prometheus-client
requests
python-jose[cryptography]
python-multipart
pyodbc
